# api/openapi.yaml
openapi: 3.1.0
info:
    title: PP Prime Construction API
    version: 0.1.0
    description: |
        สัญญา API ระหว่างหน้าบ้าน (Angular) และหลังบ้านของระบบบริหารงานก่อสร้าง PP Prime Construction

        หลักการ
        - เงินเป็นบาท (number) ทศนิยมได้ 2 ตำแหน่ง
        - วันที่ตามแผนใช้ `date` (YYYY-MM-DD) เวลาที่เกิดเหตุการณ์ใช้ `date-time` (ISO 8601 พร้อมเขตเวลา)
        - สถานะ/ประเภทเป็นรหัสภาษาอังกฤษ หน้าบ้านแปลเป็นข้อความไทยเอง
        - ข้อผิดพลาดตอบเป็น `application/problem+json` (RFC 9457)
        - การคำนวณทางธุรกิจ (ระดับผู้อนุมัติ กำไร-ขาดทุน ไทม์ไลน์ สถานะงวด) ทำที่หลังบ้าน
        - ทุกการเปลี่ยนแปลงข้อมูล หลังบ้านต้องบันทึก Audit Log เอง (หน้าบ้านไม่มี API เขียน Audit Log)
        - สิทธิ์การเข้าถึงตรวจที่หลังบ้านตามบทบาทใน `GET /roles`; ข้อมูลที่บทบาทนั้นไม่มีสิทธิ์เห็นต้องไม่ถูกส่งมา
servers:
    - url: /api
      description: ค่าเริ่มต้น (หน้าบ้านตั้ง apiBaseUrl ได้)
security:
    - bearerAuth: []
...
paths:
    /auth/login:
        post:
            tags: [auth]
            summary: เข้าสู่ระบบด้วยอีเมลและรหัสผ่าน — ได้ access token (Bearer) สำหรับทุกคำขอ
            description: |
                - อีเมลหรือรหัสผ่านไม่ถูกต้อง → 401 (ข้อความเดียวกัน ไม่บอกว่าผิดช่องไหน)
                - รหัสผ่านถูกแต่ยังเข้าใช้ไม่ได้ → 403 โดย title บอกเหตุผล: คำขอสมัครรออนุมัติ, คำขอสมัครถูกปฏิเสธ (detail = เหตุผล), บัญชีถูกระงับ, ยังไม่ได้ตั้งรหัสผ่านจากคำเชิญ
                - ผิดติดกัน 5 ครั้งต่ออีเมล → 429 ล็อก 15 นาที
                - lastLoginAt ใน user คือเวลาเข้าสู่ระบบครั้งก่อนหน้า (ก่อนครั้งนี้)
            operationId: login
            security: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/LoginInput' }
            responses:
                '200':
                    description: เข้าสู่ระบบแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/LoginResult' }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '422': { $ref: '#/components/responses/ValidationError' }
                '429':
                    description: พยายามเข้าสู่ระบบผิดหลายครั้ง
                    content:
                        application/problem+json:
                            schema: { $ref: '#/components/schemas/Problem' }
    /auth/me:
        get:
            tags: [auth]
            summary: ผู้ใช้ที่ล็อกอินอยู่
            operationId: getCurrentUser
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/CurrentUser' }
                '401': { $ref: '#/components/responses/Unauthorized' }
...
    /auth/register:
        post:
            tags: [auth]
            summary: สมัครสมาชิก (ไม่ต้องล็อกอิน) — ได้คำขอสถานะ pending เข้าระบบได้หลังแอดมินหรือเจ้าของบริษัทอนุมัติเท่านั้น
            description: |
                หลังบ้านต้องปฏิเสธการเข้าสู่ระบบของผู้ที่คำขอยังไม่ได้รับอนุมัติ
                อีเมลที่มีบัญชีอยู่แล้วหรือมีคำขอรออนุมัติอยู่ ตอบ 409
            operationId: register
            security: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/RegistrationInput' }
            responses:
                '202':
                    description: รับคำขอแล้ว รออนุมัติ
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [id, status, submittedAt]
                                properties:
                                    id: { type: string }
                                    status: { type: string, const: pending }
                                    submittedAt: { type: string, format: date-time }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /registrations:
        get:
            tags: [users]
            summary: คำขอสมัครสมาชิก (ต้องมีสิทธิ์ user.manage)
            operationId: listRegistrations
            parameters:
                - { name: status, in: query, schema: { $ref: '#/components/schemas/RegistrationStatus' } }
            responses:
                '200':
                    description: OK (ใหม่สุดก่อน)
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/Registration' }
                '403': { $ref: '#/components/responses/Forbidden' }
    /registrations/{id}/approve:
        parameters:
            - { name: id, in: path, required: true, schema: { type: string } }
        post:
            tags: [users]
            summary: อนุมัติคำขอและสร้างบัญชี (สถานะ active) — เฉพาะบทบาทแอดมินหรือเจ้าของบริษัท (403)
            description: ผูกกับบุคลากรในทะเบียน (ยังไม่มีบัญชี) และกำหนดบทบาท/สิทธิ์ตามกฎเดียวกับ POST /user-accounts คำขอที่ไม่ใช่ pending ตอบ 409
            operationId: approveRegistration
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: '#/components/schemas/UserAccountInput'
                                - type: object
                                  required: [personnelId]
                                  properties:
                                      personnelId: { type: string }
            responses:
                '200':
                    description: อนุมัติแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/UserAccount' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /registrations/{id}/reject:
        parameters:
            - { name: id, in: path, required: true, schema: { type: string } }
        post:
            tags: [users]
            summary: ปฏิเสธคำขอ (ต้องระบุเหตุผล) — เฉพาะบทบาทแอดมินหรือเจ้าของบริษัท (403)
            operationId: rejectRegistration
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [reason]
                            properties:
                                reason: { type: string, minLength: 1 }
            responses:
                '200':
                    description: ปฏิเสธแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Registration' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /users:
        get:
            tags: [auth]
            summary: รายชื่อผู้ใช้ (ใช้ในตัวกรอง)
            operationId: listUsers
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/UserRef' }
                '401': { $ref: '#/components/responses/Unauthorized' }
...
    /user-accounts:
        get:
            tags: [users]
            summary: บัญชีผู้ใช้ของบุคลากร (ต้องมีสิทธิ์ user.manage)
            operationId: listUserAccounts
            parameters:
                - { name: q, in: query, description: ค้นหาชื่อ รหัสบุคลากร อีเมล, schema: { type: string } }
                - { name: roleId, in: query, schema: { $ref: '#/components/schemas/RoleId' } }
                - { name: status, in: query, schema: { $ref: '#/components/schemas/UserAccountStatus' } }
                - { name: personnelId, in: query, description: บัญชีของบุคลากรคนนี้ (มีได้ไม่เกิน 1), schema: { type: string } }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/UserAccount' }
                '403': { $ref: '#/components/responses/Forbidden' }
        post:
            tags: [users]
            summary: สร้างบัญชีให้บุคลากรและส่งคำเชิญตั้งรหัสผ่านทางอีเมล (สถานะเริ่มต้น invited)
            operationId: createUserAccount
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/UserAccountCreateInput' }
            responses:
                '201':
                    description: สร้างแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/UserAccount' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /user-accounts/{id}:
        parameters:
            - { name: id, in: path, required: true, schema: { type: string } }
        put:
            tags: [users]
            summary: แก้บทบาท โครงการ และสิทธิ์รายคน — แก้บทบาทหรือถอน user.manage ของตัวเองไม่ได้ (409)
            operationId: updateUserAccount
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/UserAccountInput' }
            responses:
                '200':
                    description: บันทึกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/UserAccount' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /user-accounts/{id}/suspend:
        parameters:
            - { name: id, in: path, required: true, schema: { type: string } }
        post:
            tags: [users]
            summary: ระงับบัญชี (ออกจากระบบทุกอุปกรณ์ทันที) — ระงับตัวเองหรือเจ้าของบริษัทคนสุดท้ายไม่ได้ (409)
            operationId: suspendUserAccount
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                reason: { type: string, description: เหตุผล (บันทึกใน Audit Log) }
            responses:
                '200':
                    description: ระงับแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/UserAccount' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
    /user-accounts/{id}/activate:
        parameters:
            - { name: id, in: path, required: true, schema: { type: string } }
        post:
            tags: [users]
            summary: เปิดใช้บัญชีที่ระงับไว้ (กลับเป็น active หรือ invited ถ้ายังไม่เคยตั้งรหัสผ่าน)
            operationId: activateUserAccount
            responses:
                '200':
                    description: เปิดใช้แล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/UserAccount' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
    /user-accounts/{id}/reset-password:
        parameters:
            - { name: id, in: path, required: true, schema: { type: string } }
        post:
            tags: [users]
            summary: ส่งลิงก์ตั้งรหัสผ่านใหม่ทางอีเมล (บัญชี invited = ส่งคำเชิญซ้ำ) — บัญชีที่ระงับอยู่ทำไม่ได้ (409)
            operationId: resetUserPassword
            responses:
                '200':
                    description: ส่งแล้ว (invitedAt เปลี่ยนเป็นเวลาล่าสุด)
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/UserAccount' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
...
    responses:
        Unauthorized:
            description: ยังไม่ได้ล็อกอินหรือ token หมดอายุ
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
        Forbidden:
            description: บทบาทนี้ไม่มีสิทธิ์
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
        NotFound:
            description: ไม่พบข้อมูล
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
        Conflict:
            description: สถานะข้อมูลเปลี่ยนไปแล้ว เช่น คำขอถูกตัดสินไปก่อนหน้า
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
        ValidationError:
            description: ข้อมูลไม่ถูกต้อง
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
    schemas:
        Problem:
            type: object
            required: [title, status]
            properties:
                type: { type: string, format: uri, default: about:blank }
                title: { type: string, description: ข้อความสั้นสำหรับแสดงผู้ใช้ (ภาษาไทย) }
                status: { type: integer }
                detail: { type: string }
                errors:
                    type: object
                    description: ข้อผิดพลาดรายช่อง (ชื่อช่อง → ข้อความ)
                    additionalProperties: { type: string }
...
        RoleId:
            type: string
            enum: [admin, owner, admin-staff, accounting, procurement, storekeeper, auditor, project-manager, engineer, architect, foreman, safety-officer, subcontractor, client, supplier]
        UserRef:
            type: object
            required: [id, name, roleLabel]
            properties:
                id: { type: string }
                name: { type: string }
                roleLabel: { type: string, description: ชื่อบทบาท ณ เวลาที่เกิดเหตุการณ์ }
        Permission:
            type: string
            description: |
                สิทธิ์ที่หน้าบ้านใช้ตัดสินว่าจะแสดงปุ่ม/ส่วนใด (หลังบ้านต้องตรวจสิทธิ์ซ้ำทุกคำขอ)
                - progress.update: บันทึกความคืบหน้าและผลตรวจ
                - personnel.sensitive: ดู/แก้ข้อมูลอ่อนไหวของบุคลากร
                - finance.company: ดูการเงินระดับบริษัท (Dashboard กำไร-ขาดทุน)
                - approval.any: อนุมัติได้ทุกยอด
                - project.create: เปิดโครงการใหม่
                - project.manage: จัดการโครงการ เช่น บันทึกสัญญาและเริ่มแผนงาน
                - user.manage: จัดการบัญชีผู้ใช้ บทบาท และสิทธิ์ของบุคลากร
            enum: [progress.update, personnel.sensitive, finance.company, approval.any, project.create, project.manage, user.manage]
        CurrentUser:
            type: object
            required: [id, name, roleId, roleLabel, projectCodes, email, lastLoginAt, permissions]
            properties:
                id: { type: string }
                name: { type: string }
                roleId: { $ref: '#/components/schemas/RoleId' }
                roleLabel: { type: string }
                projectCodes:
                    type: array
                    description: โครงการที่เป็นสมาชิก (ว่าง = ทุกโครงการสำหรับทีมภายใน)
                    items: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                position: { type: string, description: ตำแหน่งงาน }
                department: { type: string }
                employeeCode: { type: string, description: รหัสบุคลากร (ไม่มีสำหรับบุคคลภายนอก) }
                avatarUrl: { type: string, format: uri }
                lastLoginAt: { type: string, format: date-time, description: เวลาเข้าสู่ระบบครั้งก่อนหน้า }
                permissions:
                    type: array
                    items: { $ref: '#/components/schemas/Permission' }
...
        LoginInput:
            type: object
            required: [email, password]
            properties:
                email: { type: string }
                password: { type: string, writeOnly: true }
                remember: { type: boolean, default: false, description: จดจำการเข้าสู่ระบบ (token อายุ 30 วัน แทน 8 ชั่วโมง) }
        LoginResult:
            type: object
            required: [accessToken, expiresAt, user]
            properties:
                accessToken: { type: string, description: 'ส่งใน header Authorization: Bearer <token>' }
                expiresAt: { type: string, format: date-time }
                user: { $ref: '#/components/schemas/CurrentUser' }
        RegistrationStatus:
            type: string
            enum: [pending, approved, rejected]
        RegistrationInput:
            type: object
            required: [fullName, email, phone, password]
            properties:
                fullName: { type: string, minLength: 1 }
                email: { type: string, format: email }
                phone: { type: string, description: ตัวเลข 9-10 หลัก ขึ้นต้นด้วย 0 }
                employeeCode: { type: string, description: รหัสบุคลากร (ถ้ามี) ช่วยให้ผู้อนุมัติจับคู่กับทะเบียนบุคลากรได้ }
                position: { type: string, description: ตำแหน่งงานที่แจ้งมา }
                note: { type: string, description: ข้อความถึงผู้อนุมัติ เช่น ทำงานโครงการไหน }
                password: { type: string, minLength: 8, writeOnly: true, description: ใช้ได้หลังได้รับอนุมัติ (หลังบ้านเก็บแบบ hash เท่านั้น) }
        Registration:
            type: object
            required: [id, fullName, email, phone, status, submittedAt]
            properties:
                id: { type: string }
                fullName: { type: string }
                email: { type: string }
                phone: { type: string }
                employeeCode: { type: string }
                position: { type: string }
                note: { type: string }
                status: { $ref: '#/components/schemas/RegistrationStatus' }
                submittedAt: { type: string, format: date-time }
                suggestedPersonnelId: { type: string, description: บุคลากรที่หลังบ้านจับคู่ให้ (จากรหัสบุคลากร อีเมล หรือชื่อ) และยังไม่มีบัญชี }
                decidedAt: { type: string, format: date-time }
                decidedBy: { $ref: '#/components/schemas/UserRef' }
                rejectReason: { type: string }
                accountId: { type: string, description: บัญชีที่สร้างเมื่ออนุมัติ }
        UserAccountStatus:
            type: string
            description: |
                - invited: ส่งคำเชิญแล้ว ยังไม่ได้ตั้งรหัสผ่าน
                - active: ใช้งานได้
                - suspended: ระงับการเข้าสู่ระบบ (เก็บประวัติไว้)
            enum: [invited, active, suspended]
        UserAccountInput:
            type: object
            description: |
                สิทธิ์ที่ใช้จริง = สิทธิ์ตั้งต้นของบทบาท + grantedPermissions − revokedPermissions
                บทบาทกลุ่มทีมประจำโครงการ (project) ต้องระบุ projectCodes อย่างน้อย 1 โครงการ; กลุ่มทีมภายใน (internal) เห็นทุกโครงการ ให้ส่ง projectCodes ว่าง
            required: [email, roleId, projectCodes, grantedPermissions, revokedPermissions]
            properties:
                email: { type: string, format: email, description: อีเมลที่ใช้เข้าสู่ระบบ (ซ้ำกับบัญชีอื่นไม่ได้ 409) }
                roleId: { $ref: '#/components/schemas/RoleId' }
                projectCodes: { type: array, items: { type: string } }
                grantedPermissions:
                    type: array
                    description: สิทธิ์ที่ให้เพิ่มจากบทบาท
                    items: { $ref: '#/components/schemas/Permission' }
                revokedPermissions:
                    type: array
                    description: สิทธิ์ของบทบาทที่ถอนออกสำหรับคนนี้
                    items: { $ref: '#/components/schemas/Permission' }
        UserAccountCreateInput:
            allOf:
                - $ref: '#/components/schemas/UserAccountInput'
                - type: object
                  required: [personnelId]
                  properties:
                      personnelId: { type: string, description: บุคลากรที่ยังไม่มีบัญชี (1 คน 1 บัญชี ซ้ำได้ 409) }
        UserAccount:
            allOf:
                - $ref: '#/components/schemas/UserAccountInput'
                - type: object
                  required: [id, personnelId, name, roleLabel, status, permissions, createdAt]
                  properties:
                      id: { type: string }
                      personnelId: { type: string }
                      name: { type: string, description: ชื่อจากทะเบียนบุคลากร }
                      employeeCode: { type: string }
                      position: { type: string }
                      department: { type: string }
                      phone: { type: string, description: เบอร์โทรจากทะเบียนบุคลากร }
                      roleLabel: { type: string }
                      status: { $ref: '#/components/schemas/UserAccountStatus' }
                      permissions:
                          type: array
                          description: สิทธิ์ที่ใช้จริง (หลังบ้านคำนวณ)
                          items: { $ref: '#/components/schemas/Permission' }
                      createdAt: { type: string, format: date-time }
                      invitedAt: { type: string, format: date-time, description: ส่งคำเชิญหรือลิงก์ตั้งรหัสผ่านล่าสุด }
                      lastLoginAt: { type: string, format: date-time }
...